login_page/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
psycopg[binary]==3.2.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
pydantic-settings==2.5.2
python-dotenv==1.0.1
python-multipart==0.0.12
httpx==0.28.1
email-validator
greenlet
